+++ src/components/calcite-tab-title/calcite-tab-title.tsx
@@ -75,7 +75,7 @@ export class CalciteTabTitle {
       <Host
         id={id}
         aria-controls={this.controls}
-        aria-expanded={this.isActive ? "true" : "false"}
+        aria-expanded={this.isActive.toString()}
         role="tab"
         tabindex="0"
       >