+++ stencil.config.ts
@@ -48,7 +48,8 @@ export const config: Config = {
         "calcite-date-day"
       ]
     },
-    { components: ["calcite-tree", "calcite-tree-item"] }
+    { components: ["calcite-tree", "calcite-tree-item"] },
+    { components: ["calcite-icon"] }
   ],
   outputTargets: [
     { type: "dist-hydrate-script" },
@@ -71,5 +72,10 @@ export const config: Config = {
     sass({
       injectGlobalPaths: ["src/assets/styles/includes.scss"]
     })
-  ]
+  ],
+  testing: {
+    moduleNameMapper: {
+      "^/assets/(.*)$": "<rootDir>/src/tests/iconPathDataStub.js"
+    }
+  }
 };