+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Breaking Changes
+- `calcite-progress` no longer accepts slotted content
+
 ### Added
 - `setFocus()` added to `calcite-alert` - focuses a slotted link or a close button, if present