+++ CHANGELOG.md
@@ -5,6 +5,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Breaking Changes
+- `calcite-progress` no longer accepts slotted content
+
+### Added
+- `calcite-accordion` now accepts `transparent` as an `appearance` attribute value
+- `calcite-accordion` now accepts an `icon-type` attribute to specify icon type - "chevron" (default), "caret" or "plus-minus"
+- `calcite-accordion-item` now accepts an `item-subtitle` attribute to display beneath `item-title`
+- `setFocus()` added to `calcite-alert` - focuses a slotted link or a close button, if present
+- `calcite-loader` now accepts a `no-padding` boolean attribute
+
+### Updated
+- `setFocus()` now focuses the first element in a `calcite-notice` - a slotted link or a close button, if present
+
+## [v1.0.0-beta.16] - Dec 19th 2019
+
+### Added
+- new `calcite-icon` component
+- new `CalciteModal.focusElement` method for restoring focus to an element in a modal
+- `calcite-button` now accepts boolean attributes `round` and `floating`
+- `calcite-button` can now be programmatically focused with the `setFocus()` method
+- the close button of a `dismissible` `calcite-notice` can now be programmatically focused with the `setFocus()` method
+
+### Fixed
+- fixes for date picker in Edge (#257)
+
 ## [v1.0.0-beta.15] - Nov 26th 2019
 
 ### Fixed
@@ -169,7 +195,10 @@ Fix issue with previous release.
 ## [v1.0.0-beta.1] - Aug 2nd 2019
 
 First initial beta release.
-[Unreleased]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.13...master "master"
+[Unreleased]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.16...master "master"
+[v1.0.0-beta.16]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.15...v1.0.0-beta.16 "v1.0.0-beta.16"
+[v1.0.0-beta.15]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.14...v1.0.0-beta.15 "v1.0.0-beta.15"
+[v1.0.0-beta.14]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.13...v1.0.0-beta.14 "v1.0.0-beta.14"
 [v1.0.0-beta.13]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.12...v1.0.0-beta.13 "v1.0.0-beta.13"
 [v1.0.0-beta.12]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.11...v1.0.0-beta.12 "v1.0.0-beta.12"
 [v1.0.0-beta.11]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.10...v1.0.0-beta.11 "v1.0.0-beta.11"